backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Configurazione da variabili d'ambiente
pydantic-settings>=2.5.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0

# Cache
redis>=5.0.0

# HTTP client asincrono (per chiamate ad API esterne e AI)
httpx>=0.28.0

# Flight providers
# fast-flights rimosso: non funziona da ambienti server (Google GDPR consent wall)
# GoogleFlightsProvider usa SerpAPI via httpx (già incluso sopra)
# Amadeus: provider primario (2000 req/mese, major carriers)

# Testing
pytest>=7.0
pytest-asyncio>=0.21.0
ruff>=0.9.0
